Op afrondingsfouten en een probleem met het tekenen tijdens de uitvoering werkt alles

+++ test lauren/src/domain/robots/Movement.java
@@ -20,7 +20,7 @@ public enum Movement {
 	
 	public void execute(SimRobotPilot simRobotPilot) {
 		double speed=simRobotPilot.getMovingSpeed();
-		int moveAmount = 1* speedMultiplier;
+		double moveAmount = 1* speedMultiplier;
 		int sleepTime= Math.abs((int) (1000* moveAmount/speed));
 		while(true){
 			simRobotPilot.getPosition().move(simRobotPilot.getOrientation(), moveAmount);
